[JBIDE-13158] removed connection lookup by usename and host
Removed ConnectionsModel#getConnectionByUsernameAndHost since this is
not required any more. This was used when embedding cartridges on an
existing application. This code was created when JBT was using the
legacy client. With the new client applications can be create with the
domain, no need to lookup the connection any more