[JBIDE-14342] added log message with link in case of timeout
When embedding jenkins timeouts were reached very frequently in the latest
OpenShift incarnation (Online). In these cases jenkins credentials could
not get reported. We now display an appropriate error message with a link
to the environment variables.
At the same time I now increased the default timeouts to 2 mins (was: 1 min)
in the client and included this new client in JBT/OpenShift