[JBIDE-15115] getting able to consume new message reporting
OpenShift changed the way how they report messages. What was reported in
field=result is now reported in field=null&severity=result. We need to
be able to consume the old and the new way in order to be able to talk
to OpenShift Online and existing OSE instances (their hotfix is too late
for us, we would release a non-working CR1)