[JBIDE-15309] adding markers to git when creating
The prior implementation was just creating the marker file but did not
git add it. Publishing the project to OpenShift would then not push the
new marker file since they were not added to the git index yet. I
corrected this now. 
At the same time I now also made sure the wizard is cancellable and
would only perform the operation when pushing "OK" (was:
creation/removal were immediately/on click executed).