[JBIDE-15484] Provide support for scaled app when tailing log files
Added a table in the wizard page to let the user select the gear(s) on
which the 'tail' command should be executed.
The table displays the name of the cartridges for each GearGroup along
with the number of gears that each group consumes.
The content is the table is refreshed each time the wizard is opened.

One MessageConsole per selected gear will be opened to display the 'tail'
result stream