JBIDE-20028 Import wizard: cloning and import operation is executed in modal (wizard) dialog, blocks me from doing anything.
ImportApplicationHandler is made to open a non-modal wizard dialog.
However, it should be taken into account that new/import application wizards do operations
that lock workspace resources. While the operation is going, it is ok for user just to navigate
and browse resources, but a change/create/remove action may be put to wait for the wizard's operation,
and Eclipse will be locked till its end.