[JBIDE-21429] Initial support for LiveReload
This initial support is pretty hacky. The LiveReload command requires IServerModules (that wraps an IServer and an IModule) to determine the host to proxy and the url to open
Current implementation of OpenShift server adapter doesn't contain any IModule, so we provide a stub IModule implementation, even though IModules should be created via IModuleFactories and never implemented directly.

Signed-off-by: Fred Bricon <fbricon@gmail.com>