JBIDE-20751 Wizard closing dialog shown upon canceling a creation of a project in process of creating a new OpenShift 3 application
It was the original idea to close New Application wizard entirely
if user cancelled creation of required project. There was bug in it, it is fixed.

Logic for when wizard should be closed or not is enhanced to consider these 3 cases:

1. File-> New -> OpenShift Application or JBoss Central -> OpenShift Application started;
user selects connection and after Next, Select Template page is opened
and dialog New Project is opened automatically (when project list is empty).
In this case, New Application wizard should not be closed on project creation cancel
because user has the option to push Back and select another connection.

2. Context menu on connection New -> Application is started;
dialog New Project is opened automatically (when project list is empty).
In this case New Application wizard should be closed on project creation cancel.

3. New Project dialog is started through Manage Projects link.
In this case, no matter how New Application wizard is started,
it should not be closed if project creation is cancelled.