JBIDE-22511 Application wizard: project without git remote is initially decorated with a "required" decorator, not the warning decoration (#1214)
One multi-validator may not be bound to several widgets
if each widget may have its own status message.
If change in one value may influence validation of another,
we may pass and poll all observables to all interested validators
so that they will be bound, but each validator should return status
for its own widget (bound to one of valiate observables).

The rule is: we need as many validators as we the number of status
messages that may/should be displayed at the same time. The reason
is: one validator returns one status.